Move assign_op into core/transform_common.h
