+++ core/test/CMakeLists.txt
@@ -22,7 +22,7 @@ add_executable(
   time_point_test.cpp
   value_and_variance_test.cpp
   view_index_test.cpp
-)
+  transform_common_test.cpp)
 include_directories(SYSTEM ${GMOCK_INCLUDE_DIR} ${GTEST_INCLUDE_DIR})
 target_link_libraries(
   ${TARGET_NAME}
